move GenTools to separate file
